Almost at the end of the repository part

+++ src/repository/userRepository.ts
@@ -43,6 +43,7 @@ const register = async (halfUser: HalfUserType) => {
 
         return user;
     } catch (error) {
+        console.log(error);
         throw new Error("Error registering user");
     }
 }
@@ -55,6 +56,7 @@ const findByUsername = async (username: string) => {
             "SELECT * FROM user WHERE username = ?",
             username);
     } catch (error) {
+        console.log(error);
         throw new Error("Error finding user");
     }
 
@@ -87,6 +89,7 @@ const findById = async (userId: string) => {
             "SELECT * FROM user WHERE user_id = ?",
             userId);
     } catch (error) {
+        console.log(error);
         throw new Error("Error finding user");
     }
 
@@ -112,106 +115,114 @@ const findById = async (userId: string) => {
     return user;
 }
 
-const del = async (userId: string) => {
+const del = async (userId: string, adminId: string) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "DELETE FROM user WHERE user_id = ?",
-            [userId]
+            "DELETE FROM user WHERE user_id = ? AND admin_id = ?",
+            [userId, adminId]
         );
 
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error deleting user");
     }
 };
 
 
-const updatePassword = async (userId: string, inputPassword: string) => {
+const updatePassword = async (userId: string, adminId: string, inputPassword: string) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET password = ? WHERE user_id = ?",
-            [inputPassword, userId]
+            "UPDATE user SET password = ? WHERE user_id = ? AND admin_id = ?",
+            [inputPassword, userId, adminId]
         );
 
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating password");
     }
 
 }
 
 
-const updateUsername = async (userId: string, newUsername: string) => {
+const updateUsername = async (userId: string, adminId: string, newUsername: string) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET username = ? WHERE user_id = ?",
-            [newUsername, userId]
+            "UPDATE user SET username = ? WHERE user_id = ? AND admin_id = ?",
+            [newUsername, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating username");
     }
 };
 
-const updateNickname = async (userId: string, newNickname: string | null) => {
+const updateNickname = async (userId: string, adminId: string, newNickname: string | null) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET nickname = ? WHERE user_id = ?",
-            [newNickname, userId]
+            "UPDATE user SET nickname = ? WHERE user_id = ? AND admin_id = ?",
+            [newNickname, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating nickname");
     }
 };
 
-const updateBirthday = async (userId: string, newBirthday: Date | null) => {
+const updateBirthday = async (userId: string, adminId: string, newBirthday: Date | null) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET birthday = ? WHERE user_id = ?",
-            [newBirthday, userId]
+            "UPDATE user SET birthday = ? WHERE user_id = ? AND admin_id = ?",
+            [newBirthday, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating birthday");
     }
 };
 
-const updateEmail = async (userId: string, newEmail: string | null) => {
+const updateEmail = async (userId: string, adminId: string, newEmail: string | null) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET email = ? WHERE user_id = ?",
-            [newEmail, userId]
+            "UPDATE user SET email = ? WHERE user_id = ? AND admin_id = ?",
+            [newEmail, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating email");
     }
 };
 
-const updatePhone = async (userId: string, newPhone: string | null) => {
+const updatePhone = async (userId: string, adminId: string, newPhone: string | null) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET phone = ? WHERE user_id = ?",
-            [newPhone, userId]
+            "UPDATE user SET phone = ? WHERE user_id = ? AND admin_id = ?",
+            [newPhone, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating phone");
     }
 };
 
-const updateAddress = async (userId: string, newAddress: string | null) => {
+const updateAddress = async (userId: string, adminId: string, newAddress: string | null) => {
     try {
         const [result] = await maria.query<ResultSetHeader>(
-            "UPDATE user SET address = ? WHERE user_id = ?",
-            [newAddress, userId]
+            "UPDATE user SET address = ? WHERE user_id = ? AND admin_id = ?",
+            [newAddress, userId, adminId]
         );
         return result.affectedRows === 1;
     } catch (error) {
+        console.log(error);
         throw new Error("Error updating address");
     }
 };
 
 
-export default {register, del, findById, findByUsername, updatePassword, updateAddress, updateEmail, updatePhone, updateNickname, updateBirthday}
+export default {register, del, findById, findByUsername, updatePassword, updateAddress, updateEmail, updatePhone, updateNickname, updateBirthday, updateUsername}
